+++ app/layout.tsx
@@ -2,20 +2,33 @@ import 'tailwindcss/tailwind.css';
 import localFont from '@next/font/local';
 import AnalyticsWrapper from '../components/Analytics';
 
+import styles from '../styles/palettes.module.css';
+
 const klimaWeb = localFont({
-  src: './klima-variable.ttf',
+  src: '../public/klima-variable.ttf',
   variable: '--font-klima',
 });
+
+export const revalidate = 60;
 
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
+  const paletteOptions = Object.keys(styles);
+  const palette =
+    paletteOptions[Math.floor(Math.random() * paletteOptions.length)];
   return (
     <html lang="en" className={klimaWeb.variable}>
       <body>
-        {children}
+        <div className={styles[palette]}>
+          <div className="flex justify-center w-full text-sm bg-back min-h-screen text-[1.25em] sm:text-[1.3em]">
+            <div className="flex flex-col items-start justify-center leading-9 gap-y-4 my-[4em] mx-[2em] sm:m-16 max-w-2xl text-content font-sans">
+              {children}
+            </div>
+          </div>
+        </div>
         <AnalyticsWrapper />
       </body>
     </html>
